refactor: relocate mobile stats bar component
