+++ src/components/MobileStatsBar.test.tsx
