+++ src/components/index.ts
@@ -1,11 +1,13 @@
 export { AppButton } from './AppButton';
 export { BossHealthBar } from './BossHealthBar';
+export { MobileStatsBar } from './MobileStatsBar';
 export { EncounterBrand } from './EncounterBrand';
 export { EncounterSummary } from './EncounterSummary';
 export { Modal } from './Modal';
 export { SurfaceSection } from './SurfaceSection';
 
 export type { BossHealthBarProps } from './BossHealthBar';
+export type { MobileStatsBarProps } from './MobileStatsBar';
 export type { EncounterBrandProps } from './EncounterBrand';
 export type { EncounterSummaryProps } from './EncounterSummary';
 export type { ModalProps } from './Modal';
